Share one time initialization between testing frameworks
