+++ FWCore/TestProcessor/src/oneTimeInitialization.h
@@ -1,0 +1,9 @@
+#if !defined(FWCore_TestProcessor_oneTimeInitialization_h)
+#define FWCore_TestProcessor_oneTimeInitialization_h
+
+//Used to initialize the plugin manager
+namespace edm::testprocessor {
+  bool oneTimeInitialization();
+}
+
+#endif
